refactor: extract stock transfer page components (647 to 200 lines)
- Refactored transfer page from 647 to 200 lines (69% reduction)
- Extracted components:
  - StoreSelector: From/To store selection dropdowns
  - ProductSelector: Product search with stock filtering
  - TransferSummary: Transfer overview with submit button
  - TransferHelpCard: Important notes card
  - useStockTransfer hook: Data loading and form logic
- Improved code organization and maintainability
- All form functionality preserved with cleaner architecture
